#1048 P2 (step 1): extract handle_stream into src/server.rs
First step of the main.rs → server/ subsystem split called out in
#1048's "Clean Sweep" plan. Lifts the 414-line `handle_stream`
control-socket request dispatcher out of main.rs into a dedicated
`src/server.rs`, leaving main.rs focused on bootstrap (`main`/`run`/
parse_args), the daemon-loop helpers, and the data-plane planning
fns (replan_queues, summarize_queues, etc.).

Going incremental rather than all-in-one: this PR moves only
`handle_stream` itself. The supporting helpers it calls
(refresh_status, reconcile_status_bindings, build_synced_session_*,
write_state, etc.) stay in main.rs. They're crate-private and
accessible from server.rs because server is a child module of the
binary's crate root, so child-module ancestor-privacy makes them
visible without `pub(crate)` widening.

Visibility: `fn handle_stream` widened to `pub(crate) fn handle_stream`
so main.rs's `use server::handle_stream;` resolves through the
explicit re-import.

Side cleanups:
- main.rs `use std::io::{BufRead, BufReader, BufWriter, Write};`
  collapsed (handle_stream owned all consumers; main.rs no longer
  uses any of those types).
- main.rs `use std::os::unix::net::{UnixListener, UnixStream};`
  → `use std::os::unix::net::UnixListener;` (UnixStream was only
  used by handle_stream).

LOC distribution after:
- main.rs:    872 (was 1,286 — 414 LOC moved + 2 imports trimmed)
- server.rs:  429 (new — 414-line fn + 8-line header + 7 imports)

cargo test --release: 865 passed, 0 failed.

Closes the LOC-driven portion of #1048. Subsequent PRs will move
ServerState/Args/PollMode into server/state.rs to fully realize the
user's `server/{mod,handlers,state}.rs` plan.
